azure_functions/cosmosdb/python/requirements.txt
azure-core==1.34.0
azure-cosmos==4.9.0
azure-functions==1.23.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
MarkupSafe==3.0.2
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
